Promenjen pregled rute.
Dodat bottom sheet.
