+++ app/src/main/res/drawable/ic_list_header_background.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
+
+    <solid android:color="@color/white" />
+    <corners android:topLeftRadius="16dp" android:topRightRadius="16dp" />
+
+</shape>
